move to optProb
